requirements.txt
requests
feedparser
beautifulsoup4
googletrans==4.0.0-rc1
python-dateutil
schedule
